Resolve PR comments and fix lint error.
